fastapi_backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-multipart>=0.0.12
PyYAML>=6.0.2
spacy>=3.7.0
